tests: Clarify some logic in the commandline test
The commandline test invokes fpm upon execution. It also was not
commented to document why it does what it does and the caveats of the
approach. Comments were added, along with some logic to sanitize the
environment when re-running FPM. (You cannot spawn an MPI process from
another MPI process, AFAIK, which is why failures are encountered if
you try to run the example code stand alone application if it was
built with caf.) The sanitization checks if FPM_FC is set in the
environment and removes it before calling FPM to run the example
code. If it were present it would run the version of the example built
with caf and error out. By removing FPM_FC from the environment you
ensure that a serial version is built and run.