CI: Build and run example without caf first
This ensure that the example is built and ready to go before the tests
are run. It is needed by the test and sometimes the build can throw
errors if running fpm in parallel and you don't do this (because when
running the tests in parallel, a serial library and version of the
example code is built).