test(command_line): handle missing flag values
This commit adds a test that verifies the handling of cases in
which a value associated with a command-line flag is not present.
The motivating case is supporiting a statement like

type(command_line_t) command_line
flag_value = command_line%flag_value("--foo")

when the --foo flag is missing.  This commit adds a test ensuring
that the flag_value() function resultis is then the empty string.